Test in combination with other error types
outOfOrderErrors make sense and remain fixable in combination with 
message errors.

In the presence of too many assertion arguments, something weird is 
going on, so it makes little sense to try to predict which of the 
arguments is the expected value.

outOfOrderErrors can't occur in combination with too few assertion 
arguments.